[Fix] 거래소 뷰 데이터 바인딩 오류 수정

+++ Coinone-iOS/Coinone-iOS/Source/Cells/StockTVC.swift
@@ -148,7 +148,7 @@ extension StockTVC {
 }
 
 extension StockTVC {
-    func setData(coinLogoImageName: String, coinEnglishTitle: String, coinKoreanTitle: String, coinCurrentPrice: Float, riseOrDescent: String, percentage: Float, coinTotalPrice: Float) {
+    func setData(coinLogoImageName: String, coinEnglishTitle: String, coinKoreanTitle: String, coinCurrentPrice: Float, riseOrDescent: String, percentage: Double, coinTotalPrice: Float) {
         let formatter = NumberFormatter().then {
           $0.numberStyle = .decimal
         }
